working product managment with test

+++ src/main/java/com/shop/shop/entity/ProductEntity.java
@@ -12,13 +12,13 @@ import lombok.NoArgsConstructor;
 public class ProductEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private int id;
     @Column(name = "NAME")
     private String name;
     @Column(name = "QUANTITY")
     private int quantity;
 
-    public ProductEntity(Long id,String name, int quantity) {
+    public ProductEntity(int id,String name, int quantity) {
         this.id = id;
         this.name = name;
         this.quantity = quantity;
