+++ src/main/java/com/shop/shop/model/ProductDto.java
@@ -5,11 +5,11 @@ import lombok.Getter;
 
 @Getter
 public class ProductDto {
-    private Long id;
+    private int id;
     private String name;
     private int quantity;
 
-    public ProductDto(Long id, String name, int quantity) {
+    public ProductDto(int id, String name, int quantity) {
         this.id = id;
         this.name = name;
         this.quantity = quantity;
